% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{fb_index}
\alias{fb_index}
\title{Read profile information from Facebook index.htm file}
\usage{
fb_index(file)
}
\arguments{
\item{file}{location of your index.htm file}
}
\value{
\code{data.frame} with two columns - \code{field} and \code{value}
}
\description{
Read profile information from Facebook index.htm file
}

