% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{fb_messages}
\alias{fb_messages}
\title{Read messages from Facebook html/messages.htm file}
\usage{
fb_messages(file)
}
\arguments{
\item{file}{location of your messages.htm file}
}
\value{
\code{data.frame} with 5 variables:
\describe{
 \item{thread}{thread name - participants names separated by commas}
 \item{user}{name of the message's author}
 \item{dt}{time when message was posted}
 \item{tz}{your timezone at the moment when you recieved the message}
 \item{message}{message itself}
}
}
\description{
Read messages from Facebook html/messages.htm file
}
\examples{
msgs <- fb_messages("my_fb_data/html/messages.htm")
}

