% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{fb_parse_dt}
\alias{fb_parse_dt}
\title{Parse datetime in FB data}
\usage{
fb_parse_dt(x, format = NULL)
}
\arguments{
\item{x}{character vector of datetimes}

\item{format}{datetime format id. See <link here>}
}
\description{
Parse datetime in FB data
}
\examples{
fb_parse_dt("Wednesday, 9 March 2016 at 16:01 UTC+02")
}
