% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{fb_parse_file}
\alias{fb_parse_file}
\title{Parse FB file}
\usage{
fb_parse_file(file, parser = NULL)
}
\arguments{
\item{file}{file to parse}

\item{parser}{which parser to use. This function normally should detect by itself.
But if any problems occur, you can set it by hands. Possible values are:
 "table", "ul", "thread".}
}
\description{
Parse FB file
}
\examples{
msgs <- fb_parse_file("data/html/messages.htm)
msgs <- fb_parse_file("data/html/messages.htm, parser = "thread")
}

