% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fbdata-datasets}
\alias{fbdata-datasets}
\title{Historical football data}
\source{
\url{http://www.football-data.co.uk/}
}
\description{
Football scores and match statistics for several European leagues.
}
\details{
To use a particular data set, first it must be loaded with
\code{\link{data}}.
}
\section{Datasets}{

\describe{
  \item{\code{scores.d1}}{Bundesliga scores & match statistics}
  \item{\code{scores.e0}}{Premier League scores & match statistics}
  \item{\code{scores.f1}}{Ligue 1 scores & match statistics}
  \item{\code{scores.i1}}{Serie A scores & match statistics}
  \item{\code{scores.n1}}{Eredivisie scores & match statistics}
  \item{\code{scores.p1}}{Primeira Liga scores & match statistics}
  \item{\code{scores.sp1}}{La Liga scores & match statistics}
}
}
\seealso{
\code{\link{data}}
}

