% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{fbdata-package}
\alias{fbdata-package}
\title{Historical football data}
\description{
This package includes historical football results and match
statistics for several European leagues.  It also provides a
mechanism for keeping the data up-to-date by fetching the latest
football results from
\href{http://www.football-data.co.uk/}{Football-Data.co.uk}.  (This
particular feature requires that the package's directory be
writable by the user.)
}

