% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rebuild.datasets}
\alias{rebuild.datasets}
\title{Rebuild data sets}
\usage{
rebuild.datasets()
}
\description{
Rebuilds R data files from the raw data files.
}
\details{
Data sets must be reloaded with \code{\link{data}} for any changes
to make effect.  In general, it is not necessary to call this
function after an update, because \code{\link{update.scores}}
already rebuilds any data sets that need to be rebuilt.
}
\seealso{
\code{\link{update.scores}} \code{\link{fbdata-datasets}}
}

