% Generated by roxygen2 (4.0.1): do not edit by hand
\name{update.scores}
\alias{update.scores}
\title{Update data sets}
\usage{
\method{update}{scores}(dataset = NULL, all.seasons = FALSE)
}
\arguments{
\item{dataset}{character vector.  Names of data sets to update, or
\code{NULL} to update every data set.}

\item{all.seasons}{If true, do a full update, including non-current
seasons.}
}
\description{
Updates data sets by fetching new data from the Internet.  By
default, only results from the current season are updated.  The
value of \code{getOption('current.season')} determines the current
season at run time.
}
\seealso{
\code{\link{rebuild.datasets}} \code{\link{fbdata-datasets}}
}

