% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{fb_messages}
\alias{fb_messages}
\title{Read messages from Facebook html/messages.htm file}
\usage{
fb_messages(path)
}
\arguments{
\item{path}{path to your facebook data folder or precise location of messages.htm file}
}
\value{
\code{data.frame} with 5 variables:
\describe{
 \item{\code{thread}}{thread name - participants names separated by commas}
 \item{\code{user}}{name of the message's author}
 \item{\code{dt}}{time when the message was created, in your time settings}
 \item{\code{tz}}{your timezone at the moment when you recieved the message}
 \item{\code{message}}{message itself}
}
}
\description{
Read messages from Facebook html/messages.htm file
}
\examples{
msgs <- fb_messages("my_fb_data/html/messages.htm")
}

