% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{fb_parse_dt}
\alias{fb_parse_dt}
\title{Parse datetime in FB data}
\usage{
fb_parse_dt(x, lang)
}
\arguments{
\item{x}{character vector of datetimes}

\item{lang}{language in which time is written}
}
\description{
Parse datetime in FB data
}
\examples{
# no examples currently
}
